requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-SocketIO==5.5.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
alembic==1.11.1
celery==5.3.1
redis==4.5.5
pandas==2.2.3
eventlet==0.33.3
flask-limiter==2.9.0
